+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [XX.XX.XX] - xxxx-xx-xx
+### Added
+- Added support for custom parameters feature
+
 ## [3.3.0] - 2024-01-24
 ### Added
 - Support for custom cookie header